projects/customer_churn/pyproject.toml
[project]
name = "customer-churn"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "typer>=0.12",
  "pyyaml>=6.0",
  "pydantic>=2.7",
  "dbx-platform",
  "shared-lib",
]

[project.optional-dependencies]
local = [
  "pandas>=2.2",
  "numpy>=2.0",
  "matplotlib>=3.9",
]

[project.scripts]
customer-churn = "customer_churn.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/customer_churn"]
